+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- `Spinner` added to the loading state in the profile page
+
 ## [0.57.1] - 2022-08-09
 
 ### Fixed